Move impl. of ank delete workload into own file
Issue-Id: #229
